% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census-cities.R
\docType{data}
\name{census_cities}
\alias{census_cities}
\title{Historical population data for U.S. cities}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 62224 rows and 15 columns.}
\usage{
census_cities
}
\description{
This data frame contains historical population data for U.S. cities. The data
is drawn from the U.S. Census. The data was compiled by Erik Steiner and
Jason Heppler at the Center for Spatial and Textual Analysis, Stanford
University. See their
\href{https://github.com/cestastanford/historical-us-city-populations}{description
of the data} for a fuller accounting of how the data was gathered.
}
\references{
U.S. Census Bureau and Erik Steiner, "United States Historical
  City Populations, 1790-2010," Spatial History Project, Center for Spatial
  and Textual Analysis, Stanford University,
  \url{https://github.com/cestastanford/historical-us-city-populations/}.
}
\keyword{datasets}
