% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{cb_2014_us_cd114_500k}
\alias{cb_2014_us_cd114_500k}
\alias{cb_2014_us_county_500k}
\alias{cb_2014_us_state_500k}
\alias{hist_us_counties_hires}
\alias{hist_us_states_hires}
\title{High resolution data for USAboundaries}
\format{An object of class \code{SpatialPolygonsDataFrame} with 441 rows and 9 columns.}
\usage{
cb_2014_us_cd114_500k

cb_2014_us_county_500k

cb_2014_us_state_500k

hist_us_counties_hires

hist_us_states_hires
}
\description{
The additional data in this package is documented in \code{USAboundaries}
package. See the documentation for \link[USAboundaries]{census_boundaries},
\link[USAboundaries]{hist_us_states}, and
\link[USAboundaries]{hist_us_counties}.
}
\keyword{datasets}

